В ресурсах модификатор login_required заменён на вызов функции проверки авторизации

+++ resources/article_likes.py
@@ -1,20 +1,21 @@
 from flask import jsonify
 from flask_restful import abort as fr_abort, Resource
-from flask_login import current_user, login_required
+from flask_login import current_user
 from model_workers.article_like import ArticleLikeModelWorker
 from tools.errors import ArticleNotFoundError, LikeNotFoundError, LikeAlreadyThereError
+from tools.check_authorization import check_authorization
 
 
 class ArticleLikeResource(Resource):
-    @login_required
     def get(self, article_id):
+        check_authorization()
         return jsonify({"like_exist": ArticleLikeModelWorker.like_exist({
             "article_id": article_id,
             "user_id": current_user.id
         })})
 
-    @login_required
     def post(self, article_id):
+        check_authorization()
         try:
             ArticleLikeModelWorker.new_like({
                 "article_id": article_id,
@@ -27,8 +28,8 @@ class ArticleLikeResource(Resource):
         else:
             return jsonify({"success": "ok"})
 
-    @login_required
     def delete(self, article_id):
+        check_authorization()
         try:
             ArticleLikeModelWorker.delete_like({
                 "article_id": article_id,
@@ -37,6 +38,6 @@ class ArticleLikeResource(Resource):
         except ArticleNotFoundError:
             fr_abort(404, message=f"Article {article_id} not found")
         except LikeNotFoundError:
-            fr_abort(400, message="Like not found")
+            fr_abort(404, message="Like not found")
         else:
             return jsonify({"success": "ok"})
