+++ resources/comments.py
@@ -1,11 +1,12 @@
 from flask import jsonify
 from flask_restful import abort as fr_abort, Resource
-from flask_login import current_user, login_required
+from flask_login import current_user
 from parsers import get_comment_parser, add_comment_parser, put_comment_parser
 from tools.image_to_byte_array import image_to_byte_array
 from tools.hex_image_to_file_storage import hex_image_to_file_storage
 from tools.constants import COMMENTS_IMAGES_DIR
 from tools.errors import CommentNotFoundError, ArticleNotFoundError, ForbiddenToUserError
+from tools.check_authorization import check_authorization
 from model_workers.comment import CommentModelWorker
 
 
@@ -24,9 +25,9 @@ class CommentResource(Resource):
                     ).hex()
             return jsonify({"comment": comment})
 
-    @login_required
     def put(self, comment_id):
         args = put_comment_parser.parser.parse_args()
+        check_authorization()
         comment_data = {"text": args["text"]}
         if args.get("image") is not None:
             comment_data["image"] = hex_image_to_file_storage(args["image"])
@@ -39,8 +40,8 @@ class CommentResource(Resource):
         else:
             return jsonify({"success": "ok"})
 
-    @login_required
     def delete(self, comment_id):
+        check_authorization()
         try:
             CommentModelWorker.delete_comment(comment_id, current_user.id)
         except CommentNotFoundError:
@@ -65,9 +66,9 @@ class CommentsListResource(Resource):
                     ).hex()
         return jsonify({"comments": comments})
 
-    @login_required
     def post(self):
         args = add_comment_parser.parser.parse_args()
+        check_authorization()
         comment_data = {
             "author": current_user.id,
             "article_id": args["article_id"],
